[refactor] Remove response and file logic from cli.py
